+++ docs/whats_new.rst
@@ -1,5 +1,10 @@
 .. _WhatsNewAnchor:
 
+What's new in version 2.0.3 (2019-04-10)
+========================================
+
+* Fixes an issue when requesting an unknown controller method.
+
 What's new in version 2.0.2 (2019-01-23)
 ========================================